adding new task to the connected user feature

+++ src/main/java/ma/emsi/todo_pfa/controller/TaskController.java
@@ -24,7 +24,7 @@ public class TaskController {
 	@Autowired
 	TaskService taskSer;
 	
-	public long getCurrentUserId() {
+	public int getCurrentUserId() {
 		return 1;
 	}
