+++ src/main/java/ma/emsi/todo_pfa/service/TaskService.java
@@ -2,12 +2,14 @@ package ma.emsi.todo_pfa.service;
 
 import java.util.List;
 
+import org.springframework.stereotype.Service;
+
 import ma.emsi.todo_pfa.entity.Task;
 
 public interface TaskService {
 
-	Task add(Task task, long userid);
+	Task add(Task task, int userid);
 
-	List<Task> getUserTasks(long userId);
+	List<Task> getUserTasks(int userId);
 
 }
